Move public_erp.py to provider module.
