+++ erp/views.py
@@ -35,9 +35,8 @@ from .forms import (
     ViewAccessibiliteForm,
 )
 from .models import Accessibilite, Activite, Commune, Erp, Vote
-from .provider import sirene
+from .provider import public_erp, sirene
 from . import naf
-from . import public_erp
 from . import schema
 from . import serializers
 from . import versioning
